feat: view 분리
